Rename the fetchSourceChannels and carry it to utils
